# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Master API
  description: |
    The Task Master API provides comprehensive project and task management capabilities with multi-tenant organization support.

    ## Features
    - Multi-tenant architecture with organization management
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Comprehensive task and project management
    - Real-time capabilities with webhooks
    - Advanced filtering and search

    ## Authentication
    All endpoints require authentication using Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```

    ## Rate Limiting
    - Authentication endpoints: 10 requests per minute per IP
    - General API endpoints: 1000 requests per hour per user
    - Bulk operations: 100 requests per hour per user
  version: 1.0.0
  contact:
    name: Task Master API Support
    email: api-support@taskmaster.com
    url: https://docs.taskmaster.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.taskmaster.com/v1
    description: Production server
  - url: http://localhost:3001/api/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authentication
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100

    # User and authentication schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        email:
          type: string
          format: email
          example: 'john.doe@example.com'
        fullName:
          type: string
          example: 'John Doe'
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: 'https://example.com/avatar.jpg'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-10T00:00:00Z'
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '950e8400-e29b-41d4-a716-446655440004'
        projectId:
          type: string
          format: uuid
          example: '850e8400-e29b-41d4-a716-446655440003'
        title:
          type: string
          example: 'Design homepage mockup'
        description:
          type: string
          nullable: true
          example: 'Create wireframes and high-fidelity mockups'
        status:
          type: string
          enum: [todo, in_progress, completed, cancelled]
          example: 'todo'
        priority:
          type: string
          enum: [low, medium, high]
          example: 'high'
        assigneeId:
          type: string
          format: uuid
          nullable: true
          example: '550e8400-e29b-41d4-a716-446655440000'
        dueDate:
          type: string
          format: date-time
          nullable: true
          example: '2024-02-15T17:00:00Z'
        estimatedHours:
          type: number
          format: float
          nullable: true
          example: 8.0
        actualHours:
          type: number
          format: float
          nullable: true
          example: 7.5
        tags:
          type: array
          items:
            type: string
          example: ['design', 'homepage', 'ui']
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-05T10:00:00Z'
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: '2024-01-15T16:30:00Z'
    TaskWithAssignee:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            assignee:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                  format: uuid
                fullName:
                  type: string
                avatarUrl:
                  type: string
                  format: uri
                  nullable: true
            subtaskCount:
              type: integer
              example: 3
            completedSubtaskCount:
              type: integer
              example: 1
    Subtask:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 'a50e8400-e29b-41d4-a716-446655440005'
        taskId:
          type: string
          format: uuid
          example: '950e8400-e29b-41d4-a716-446655440004'
        title:
          type: string
          example: 'Create wireframes'
        description:
          type: string
          nullable: true
          example: 'Design basic wireframes for layout'
        status:
          type: string
          enum: [todo, completed]
          example: 'todo'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-03T14:00:00Z'
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: '2024-01-03T14:00:00Z'

    # Request schemas
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          example: 'Design homepage mockup - Updated'
        description:
          type: string
          maxLength: 2000
          example: 'Updated description'
        status:
          type: string
          enum: [todo, in_progress, completed, cancelled]
          example: 'completed'
        priority:
          type: string
          enum: [low, medium, high]
          example: 'medium'
        assigneeId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        dueDate:
          type: string
          format: date-time
          example: '2024-02-20T17:00:00Z'
        estimatedHours:
          type: number
          format: float
          minimum: 0
          example: 8.0
        actualHours:
          type: number
          format: float
          minimum: 0
          example: 7.5
        tags:
          type: array
          items:
            type: string
          example: ['design', 'homepage', 'ui', 'completed']
    CreateSubtaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          example: 'Create wireframes'
        description:
          type: string
          maxLength: 1000
          example: 'Design basic wireframes for layout'
...
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Task ID
...
paths:
...
  /users/profile:
    get:
      tags:
        - User Profile
      summary: Get user profile
      description: Get the current user's profile information.
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          profile:
                            $ref: '#/components/schemas/User'

    put:
      tags:
        - User Profile
      summary: Update user profile
      description: Update the current user's profile information.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                  maxLength: 100
                  example: 'John Smith'
                avatarUrl:
                  type: string
                  format: uri
                  example: 'https://example.com/new-avatar.jpg'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          profile:
                            $ref: '#/components/schemas/User'
  /users/password:
    put:
      tags:
        - User Profile
      summary: Change password
      description: Change the current user's password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  example: 'CurrentPassword123!'
                newPassword:
                  type: string
                  minLength: 8
                  example: 'NewPassword123!'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: 'Password updated successfully'

  # Organization endpoints
...
  /tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get task
      description: Get detailed information about a specific task.
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          task:
                            $ref: '#/components/schemas/TaskWithAssignee'
                          subtasks:
                            type: array
                            items:
                              $ref: '#/components/schemas/Subtask'
                          dependencies:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: string
                                  format: uuid
                                title:
                                  type: string
                                status:
                                  type: string

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update task information.
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          task:
                            $ref: '#/components/schemas/Task'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a task.
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: 'Task has been deleted successfully'

  # Subtask endpoints
  /tasks/{taskId}/subtasks:
    post:
      tags:
        - Subtasks
      summary: Add subtask
      description: Add a subtask to a task.
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubtaskRequest'
      responses:
        '201':
          description: Subtask created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          subtask:
                            $ref: '#/components/schemas/Subtask'
...